pyproject.toml
[tool.poetry]
name = "robot-llm"
version = "0.1.0"
description = ""
authors = ["Jacob Sansbury <sansburyjacob@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dateutil = "^2.9.0.post0"
bson = "^0.5.10"
strenum = "^0.4.15"
openai = "^1.30.1"
lark = "^1.1.9"
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
